// src/redux/enums/CategoriaEnum.ts
export enum CategoriaEnum {
   POPULATE_CATEGORIA_START = 'POPULATE_CATEGORIA_START',
   POPULATE_CATEGORIA_SUCCESS = 'POPULATE_CATEGORIA_SUCCESS',
   POPULATE_CATEGORIA_ERROR = 'POPULATE_CATEGORIA_ERROR',

   ADD_CATEGORIA_START = 'ADD_CATEGORIA_START',
   ADD_CATEGORIA_SUCCESS = 'ADD_CATEGORIA_SUCCESS',
   ADD_CATEGORIA_ERROR = 'ADD_CATEGORIA_ERROR',

   UPDATE_CATEGORIA_START = 'UPDATE_CATEGORIA_START',
   UPDATE_CATEGORIA_SUCCESS = 'UPDATE_CATEGORIA_SUCCESS',
   UPDATE_CATEGORIA_ERROR = 'UPDATE_CATEGORIA_ERROR',

   DELETE_CATEGORIA_START = 'DELETE_CATEGORIA_START',
   DELETE_CATEGORIA_SUCCESS = 'DELETE_CATEGORIA_SUCCESS',
   DELETE_CATEGORIA_ERROR = 'DELETE_CATEGORIA_ERROR'
}